backend/employer-admin/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1

motor==3.6.0
beanie==1.27.0

pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1

python-dotenv==1.0.1

firebase-admin==6.5.0
sentence-transformers==2.5.1
transformers==4.44.2   
faiss-cpu>=1.8.0
torch>=2.1.0
numpy>=1.24.3
httpx>=0.27.0